fix:  content_type compress br
#398
